refact(resources/**/*.js) resources reorganised
