+++ src/classes/resources/patients/index.js
@@ -1,5 +1,5 @@
 // @flow
-import Common from '../common';
+import Common from '../../common';
 
 export default class Patients extends Common {
   /**
