+++ src/classes/resources/specialities/index.js
@@ -1,5 +1,5 @@
 // @flow
-import Common from '../common';
+import Common from '../../common';
 
 export default class Specialities extends Common {
   /**
